Make linting work again with elm-analyse 0.14.2
The communication is now updated to the new message format. Because it provides the description so the hardcoded warning list got removed. Also it does not stop anymore on the first error and gives a list of all message parsing errors. The errors gives a hint that we expect at least the version 0.14.2.