Fix for Elm 0.19, find right project directory
I found this bug because I had an `elm-package.json` in my `~` directory, so when it tried to find the project dir for my Elm 0.19 project, it traversed upwards to my home dir, and then started indexing all files in my home directory for modules. This lead to the editor being unresponsive for a number of seconds.